refactor - move widget logic outside of UI comp
